+++ apps/src/tests/issue-tests/index.ts
@@ -28,7 +28,6 @@ export { default as Test706 } from './Test706'; // [E2E skipped]: can't check fo
 export { default as Test726 } from './Test726'; // [E2E created](iOS): issue related to iOS
 export { default as Test748 } from './Test748'; // [E2E skipped]: issue unclear, Test432 covers more complex cases already
 export { default as Test750 } from './Test750'; // [E2E skipped]: can't check insets
-export { default as Test755 } from './Test755';
 export { default as Test758 } from './Test758'; // [E2E created](iOS): PR related to iOS search bar
 export { default as Test761 } from './Test761'; // [E2E skipped]: can't check animation in a meaningful way
 export { default as Test779 } from './Test779'; // [E2E skipped]: can't check animation in a meaningful way
@@ -208,6 +207,7 @@ export { default as Test4276 } from './Test4276';
 export { default as Test4351 } from './Test4351';
 export { default as Test4357 } from './Test4357';
 export { default as Test4361 } from './Test4361';
+export { default as Test4423 } from './Test4423';
 export { default as TestScreenAnimation } from './TestScreenAnimation';
 // The following test was meant to demo the "go back" gesture using Reanimated
 // but the associated PR in react-navigation is currently put on hold
